+++ test/parameter/types/date_test.exs
@@ -1,0 +1,4 @@
+defmodule Parameter.Types.DateTest do
+  use ExUnit.Case
+  doctest Parameter.Types.Date
+end