+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v0.6.x (2022-11-02)
+
+### Enhancements
+
+  * [Parameter] API changes to support new [parameter_ecto](https://github.com/phcurado/parameter_ecto) library.
+  * [Parameter] Errors when parsing list return as `{index, reason}` now instead of `{:#{index}, reason}` to avoid atoms creation.
+  * [Parameter.Enum] Deprecated `as` in favour of `key`
+
 ## v0.5.x (2022-10-26)
 
 ### Enhancements